app/requirements.txt
fastapi
uvicorn
firebase-admin
google-cloud-firestore
python-dotenv
requests
pytest
gunicorn
pydantic
typing